feat: Implement client configuration management for Windows, Mac, and Linux in MainWindowEditor
