+++ Assets/root/Editor/Scripts/UI/Window/MainWindowEditor.ClientConfigure.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fcb07454b9722204b86cc2bcc39f444a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
