+++ Assets/root/Editor/Scripts/UI/Window/MainWindowEditor.CreateGUI.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Linq;
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using com.IvanMurzak.Unity.MCP.Common;
@@ -143,136 +144,6 @@ namespace com.IvanMurzak.Unity.MCP.Editor
 
             var rawJsonField = root.Query<TextField>("rawJsonConfiguration").First();
             rawJsonField.value = Startup.RawJsonConfiguration;
-        }
-
-        void ConfigureClientsWindows(VisualElement root)
-        {
-            ConfigureClient(root.Query<VisualElement>("ConfigureClient-Claude").First(),
-                configPath: Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData),
-                    "Claude",
-                    "claude_desktop_config.json"
-                ));
-
-            ConfigureClient(root.Query<VisualElement>("ConfigureClient-VS-Code").First(),
-                configPath: "Unknonwn");
-
-            ConfigureClient(root.Query<VisualElement>("ConfigureClient-Cursor").First(),
-                configPath: Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.UserProfile),
-                    ".cursor",
-                    "mcp.json"
-                ));
-        }
-
-        void ConfigureClientsMacAndLinux(VisualElement root)
-        {
-            ConfigureClient(root.Query<VisualElement>("ConfigureClient-Claude").First(),
-                configPath: Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.UserProfile),
-                    "Library",
-                    "Application Support",
-                    "Claude",
-                    "claude_desktop_config.json"
-                ));
-
-            ConfigureClient(root.Query<VisualElement>("ConfigureClient-VS-Code").First(),
-                configPath: "Unknonwn");
-
-            ConfigureClient(root.Query<VisualElement>("ConfigureClient-Cursor").First(),
-                configPath: Path.Combine(
-                    Environment.GetFolderPath(Environment.SpecialFolder.UserProfile),
-                    ".cursor",
-                    "mcp.json"
-                ));
-        }
-
-        void ConfigureClient(VisualElement root, string configPath)
-        {
-            var statusCircle = root.Query<VisualElement>("configureStatusCircle").First();
-            var statusText = root.Query<Label>("configureStatusText").First();
-            var btnConfigure = root.Query<Button>("btnConfigure").First();
-
-            var isConfiguredResult = IsMcpClientConfigured(configPath);
-
-            statusCircle.RemoveFromClassList(USS_IndicatorClass_Connected);
-            statusCircle.RemoveFromClassList(USS_IndicatorClass_Connecting);
-            statusCircle.RemoveFromClassList(USS_IndicatorClass_Disconnected);
-
-            statusCircle.AddToClassList(isConfiguredResult
-                ? USS_IndicatorClass_Connected
-                : USS_IndicatorClass_Disconnected);
-
-            statusText.text = isConfiguredResult ? "Configured" : "Not Configured";
-            btnConfigure.text = isConfiguredResult ? "Reconfigure" : "Configure";
-
-            btnConfigure.RegisterCallback<ClickEvent>(evt =>
-            {
-                var configureResult = ConfigureMcpClient(configPath);
-
-                statusText.text = configureResult ? "Configured" : "Not Configured";
-
-                statusCircle.RemoveFromClassList(USS_IndicatorClass_Connected);
-                statusCircle.RemoveFromClassList(USS_IndicatorClass_Connecting);
-                statusCircle.RemoveFromClassList(USS_IndicatorClass_Disconnected);
-
-                statusCircle.AddToClassList(configureResult
-                    ? USS_IndicatorClass_Connected
-                    : USS_IndicatorClass_Disconnected);
-
-                btnConfigure.text = configureResult ? "Reconfigure" : "Configure";
-            });
-        }
-
-        bool IsMcpClientConfigured(string configPath)
-        {
-            if (string.IsNullOrEmpty(configPath) || !File.Exists(configPath))
-                return false;
-
-            try
-            {
-                var json = File.ReadAllText(configPath);
-                return json.Contains(Startup.ServerExecutableFile);
-            }
-            catch (Exception ex)
-            {
-                Debug.LogError($"Error reading config file: {ex.Message}");
-                return false;
-            }
-        }
-        bool ConfigureMcpClient(string configPath)
-        {
-            if (string.IsNullOrEmpty(configPath) || !File.Exists(configPath))
-                return false;
-
-            try
-            {
-                var json = File.ReadAllText(configPath);
-
-                // Parse the existing config as JsonObject
-                var rootObj = JsonNode.Parse(json)?.AsObject();
-                if (rootObj == null)
-                    throw new Exception("Config file is not a valid JSON object.");
-
-                // Parse the injected config as JsonObject
-                var injectObj = JsonNode.Parse(Startup.RawJsonConfiguration)?.AsObject();
-                if (injectObj == null)
-                    throw new Exception("Injected config is not a valid JSON object.");
-
-                // Merge: add/overwrite all properties from injectObj into rootObj
-                foreach (var kv in injectObj)
-                    rootObj[kv.Key] = kv.Value;
-
-                // Write back to file
-                File.WriteAllText(configPath, rootObj.ToJsonString(new JsonSerializerOptions { WriteIndented = true }));
-
-                return IsMcpClientConfigured(configPath);
-            }
-            catch (Exception ex)
-            {
-                Debug.LogError($"Error reading config file: {ex.Message}");
-                return false;
-            }
         }
     }
 }
